index: create module moving code from ls-files
