+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod fs;
+pub mod index;
 pub mod sub_commands;
